requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
celery==5.2.7
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
Django==3.2.6
django-cors-headers==4.3.1
django_neomodel==0.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
idna==3.6
inflection==0.5.1
jexp==0.1.2
lucene-querybuilder==0.2
mysqlclient==2.2.1
neo4django==0.1.7
neo4j==5.15.0
neo4jrestclient==2.1.1
neomodel==5.2.1
openai==0.28.0
packaging==23.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
